Dairy Foods Information on all dairy plants in the United States

Dairy industry directory of dairy processors that supply milk, milk-based beverages, cheese, yogurt, ice cream, and other dairy foods and beverages

Usage

remotes::install_github("nBhabish/dairyplantdata")
library(dairyplantdata)
dairy_plant_data

Format

A tibble with 1,983 rows and 35 variables:

company_name

chr Name of the Dairy Organization

company_type

chr Products the Dairy Organization specialized in

primary_product1

chr denoting first primary product

primary_product2

chr denoting second primary product

primary_product3

chr denoting third primary product

primary_product4

chr denoting fourth primary product

primary_product5

chr denoting fifth primary product

primary_product6

chr denoting sixth primary product

primary_product7

chr denoting seventh primary product

primary_product8

chr denoting eighth primary product

primary_product9

chr denoting ninth primary product

primary_product10

chr denoting tenth primary product

primary_product11

chr denoting eleventh primary product

primary_product12

chr denoting twelfth primary product

primary_product13

chr denoting thirteenth primary product

primary_product14

chr denoting fourteenth primary product

primary_product15

chr denoting fifteenth primary product

primary_product16

chr denoting sixteenth primary product

primary_product17

chr denoting seventeenth primary product

primary_product18

chr denoting eighteenth primary product

primary_product19

chr denoting nineteenth primary product

primary_product20

chr denoting twentieth primary product

primary_product21

chr denoting twenty-first primary product

primary_product22

chr denoting twenty-second primary product

primary_product23

chr denoting twenty-third primary product

primary_product24

chr denoting twenty-fourth primary product

primary_product25

chr denoting twenty-fifth primary product

primary_product26

chr denoting twenty-sixth primary product

primary_product27

chr denoting twenty-seventh primary product

primary_product28

chr denoting twenty-eighth primary product

primary_product29

chr denoting twenty-ninth primary product

primary_product30

chr denoting thirty primary product

primary_product31

chr denoting thirty-first primary product

primary_product32

chr denoting thirty-second primary product

primary_product33

chr denoting thirty-third primary product

Source

https://www.dairyfoods.com/directories